+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## RudeHash 5.1 (TBD)
 
+* Add support for Zergpool
 * Add support for Bitcore, BitSend
 * Add support for TimeTravel10, Xevan
 * Add support for ccminer-xevan
+* Don't require a restart when changing the region
 
 ## RudeHash 5.0 (2018-03-09)